+++ src/BayesSizeAndShape.jl
@@ -1,24 +1,14 @@
 module BayesSizeAndShape
  
-
-#using Pkg
-#Pkg.activate("/Users/gianlucamastrantonio/Dropbox (Politecnico di Torino Staff)/lavori/gitrepo/BayesSizeAndShape")
-
 ##### Packages
-# using MKL
 using Distributions, Random
 using LinearAlgebra, PDMats
 using ProgressMeter
 using Kronecker
 using DataFrames
+using CategoricalArrays
 using StatsModels
 using ToggleableAsserts
-# using ThreadsX, SparseArrays
-# using ThreadedSparseArrays
-# using SparseMatricesCSR
-# using ThreadedSparseCSR
-# # using Impute
-# # using SpecialFunctions
 
 ##### Include
 include(joinpath("types.jl"))
@@ -27,24 +17,6 @@ include(joinpath("mcmc.jl"))
 include(joinpath("sampler.jl"))
 include(joinpath("wrappers.jl"))
 include(joinpath("deprecated.jl"))
-
-# include(joinpath("func_covariance_new.jl"))
-# include(joinpath("func_project.jl"))
-# include(joinpath("new_func_project.jl"))
-# include(joinpath("func_mcmc.jl"))
-
-
-# include(joinpath("func_adapt.jl"))
-# include(joinpath("sampler.jl"))
-# include(joinpath("sampler_new.jl"))
-
-# # include(joinpath("sampler_notraset.jl")) 
-# # include(joinpath("func_adapt_notraset.jl")) 
-
-
-# include(joinpath("predictw.jl"))
-
-
 
 ##### Functions
  export
@@ -55,10 +27,10 @@ include(joinpath("deprecated.jl"))
     DoKeepReflection,
     GeneralSigma,
     standardize_reg,
-    SizeAndShapeMCMC_p2withreflection,
+    SizeAndShapeMCMC,
     Valuep2,
     compute_designmatrix,
-    compute_dessignmatrix
+    compute_dessignmatrix,
+    compute_helmertized_configuration
 
-
-end # module
+end