+++ src/wrappers.jl
@@ -1,6 +1,5 @@
-
 """
-    SizeAndShapeMCMC_p2withreflection(;
+    SizeAndShapeMCMC(;
         dataset::Array{Float64,3}, 
         fm::FormulaTerm = @formula(1~ 1),
         covariates::DataFrame,
@@ -13,10 +12,12 @@
         sigmaprior::ContinuousMatrixDistribution,
         beta_init::Matrix{Float64},
         sigma_init::Symmetric{Float64,Matrix{Float64}},
-        rmat_init::Array{Float64,3}
-)
-Posterior sample from the size-and-shape model for two-dimensional data with reflection information. 
-The functions returns 4 arrays, with the psoterior samples of the regressive coefficients, covariance matrix, rotation matrices, and its rotation angles, respectively
+        rmat_init::Array{Float64,3},
+        reflection::Bool = true
+        )
+
+Posterior samples from the size-and-shape model - in this version, only two-dimensional data with reflection information are allowed. 
+The functions returns 2 Dataframes, with the posterior samples of the regressive coefficients and elements of the covariance matrix.
 
 
 # Arguments
@@ -31,15 +32,16 @@ The arguments are
 
 - `dataset::Array{Float64,3}`: Array with the data - dimension (k,p,n) of size-and-shape data. Use the function `compute_ss_from_pre` to obtain the size-and-shape data from pre-forms
 - `fm::FormulaTerm = @formula(1~ 1)`: a formula that specifies the model - the left-and size should be 1
-- `covariates::DataFrame`: a DataFrame containing the covariates - dimension (n,d). The names used in `fm` must be column names of  `covariates`
+- `covariates::DataFrame`: a DataFrame containing the covariates - dimension (n,d). The names used in `fm` must be column names of  `covariates`. Only Real And Factor covariates are allowed. The numeric column are standardized internally.
 - `iterations::NamedTuple{(:iter, :burnin, :thin),Tuple{Int64,Int64,Int64}}`: values of the iterations (iter), thin and burnin of the MCMC algorithm
 - `betaprior::ContinuousUnivariateDistribution`: The prior on the regressive coefficients - only a Normal distribution is allowed
 - `sigmaprior::ContinuousMatrixDistribution`: The prior on the covariance matrix - only an Inverse Wishart is allowed
 - `beta_init::Matrix{Float64}`: initial values for the regressive coefficients - dimension (k*d,p) 
 - `sigma_init::Symmetric{Float64,Matrix{Float64}}`: initial values for the covariance matrix - dimension (k,k) (it must be a valid covariance matrix) 
 - `rmat_init::Array{Float64,3}`: initial values for the rotation matrices - dimension (p,p,n) (each [1:p, 1:p, i], i = 1,2,...,n, must be a valid rotation matrix) 
+- `reflection::Bool`: true for a model with reflection information.
 """
-function SizeAndShapeMCMC_p2withreflection(;
+function SizeAndShapeMCMC(;
     dataset::Array{Float64,3}, 
     fm::FormulaTerm = @formula(1~ 1),
     covariates::DataFrame, #
@@ -52,10 +54,18 @@ function SizeAndShapeMCMC_p2withreflection(;
     sigmaprior::ContinuousMatrixDistribution,
     beta_init::Matrix{Float64},
     sigma_init::Symmetric{Float64,Matrix{Float64}},
-    rmat_init::Array{Float64,3}
+    rmat_init::Array{Float64,3},
+    reflection::Bool = true,
 )
 
-return mcmc(; 
+    p::Int64 = size(dataset, 2)
+    if p != 2
+        error("the current implementation allows only `size(dataset, 2) = 2 (2-dimensional data)`")
+    end
+    if reflection == false
+        error("the current implementation allows only `reflection = true`")
+    else
+         betaout, sigmaout, _, _,colnames_modelmatrix,  kout, pout, nout, dout =generalSizeAndShapeMCMC(; 
             dataset = dataset,
             fm = fm,
             covariates = covariates,
@@ -68,6 +78,17 @@ return mcmc(;
             dormat = true,
             reflection = KeepReflection(),
             sigmatype = GeneralSigma()
-)
+        )
+        nsim = size(betaout,1)
+        bbb = DataFrame(reshape(betaout, nsim, kout*dout*pout ), :auto)
+        rename!(bbb,repeat(colnames_modelmatrix,inner = kout, outer= pout) .* "| mark:" .* string.(repeat(1:kout, outer = dout*pout)) .* "| dim:" .* string.(repeat(1:pout, inner = dout*kout)))
+
+
+        sss = DataFrame(reshape(sigmaout, nsim, kout*kout ), :auto)
+        rename!(sss,"s_".* string.(repeat(1:kout, inner = kout)) .* "," .* string.(repeat(1:kout, outer = kout))  )
+
+        return bbb, sss
+    end
+    
 
 end