+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [v0.1.4] - 2023-03-14
+### Added
+- compute_helmertized_configuration
+
+### Changed
+- SizeAndShapeMCM - the output are now DataFrames
+- SizeAndShapeMCMC_p2withreflection is now SizeAndShapeMCMC
+
+
 ## [v0.1.3] - 2023-03-14
 ### Added
 - SizeAndShapeMCMC_p2withreflection() - MCMC samples for two-dimensional data with reflection information